user_service/pyproject.toml
[tool.poetry]
name = "db-services"
version = "0.1.0"
description = ""
authors = ["sabilaakram <139713987+sabilaakram@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
uvicorn = "^0.30.0"
pytest = "^8.2.1"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg = {extras = ["binary"], version = "^3.1.19"}
passlib = {extras = ["brcypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
